+++ src/sf/Random.cpp
@@ -6,9 +6,9 @@ uint32_t Random::nextU32()
 {
 	uint64_t oldstate = state;
 	state = oldstate * 6364136223846793005ULL + inc;
-	uint32_t xorshifted = ((oldstate >> 18u) ^ oldstate) >> 27u;
+	uint32_t xorshifted = (uint32_t)(((oldstate >> 18u) ^ oldstate) >> 27u);
 	uint32_t rot = oldstate >> 59u;
-	return (xorshifted >> rot) | (xorshifted << ((-rot) & 31));
+	return (xorshifted >> rot) | (xorshifted << (((uint32_t)-(int32_t)rot) & 31));
 }
 
 float Random::nextFloat()
